% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_shiny_related.R
\name{anem_app}
\alias{anem_app}
\title{Run anem shiny app}
\usage{
anem_app(shiny_app = "anem-app")
}
\description{
Run anem-app
}
\details{
Run anem-app hiny application
}
\examples{
library(anem)
\dontrun{
anem_app()
}
}
