% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_data.R
\docType{data}
\name{anem_app_scenario}
\alias{anem_app_scenario}
\title{anem-app scenario}
\format{a list object containing named items associated with anem-app}
\usage{
anem_app_scenario
}
\description{
anem-app scenario
}
\keyword{internal}
