% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_data.R
\docType{data}
\name{aquifer_confined_example}
\alias{aquifer_confined_example}
\title{Aquifer confined example}
\format{aquifer object set as \code{aquifer_type="confined"}}
\usage{
aquifer_confined_example
}
\description{
Aquifer confined example
}
\keyword{datasets}
