% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_data.R
\docType{data}
\name{aquifer_unconfined_example}
\alias{aquifer_unconfined_example}
\title{Aquifer unconfined example}
\format{aquifer object set as \code{aquifer_type="unconfined"}}
\usage{
aquifer_unconfined_example
}
\description{
Aquifer unconfined example
}
\keyword{datasets}
