% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_data.R
\docType{data}
\name{bounds_example}
\alias{bounds_example}
\title{Bounds example}
\format{bounds \code{data.frame} object (needs to be supplyed to \code{define_bounds})}
\usage{
bounds_example
}
\description{
Bounds example
}
\keyword{internal}
