% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_geoprocessing.R
\name{bounds_to_sf2}
\alias{bounds_to_sf2}
\title{Bounds to SF 2}
\usage{
bounds_to_sf2(bounds, crs)
}
\arguments{
\item{bounds}{A data.frame containing x1, y1, x2, y2, bID}

\item{crs}{crs object for sf package}
}
\value{
returns object
}
\description{
Get sf object from boundaries
}
\examples{
\dontrun{
bounds <- define_bounds(data.frame(m=c(1,-1,1,-1),b=c(0,2,2,4),bound_type=c("CH","NF","NF","NF")))
bounds_sf <- bounds_to_sf(bounds, crs=4326)
bounds_sf <- use_anem_function("bounds_to_sf",bounds=bounds,crs=4326)
}
}
\keyword{internal}
