% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_assess.R
\name{check_aquifer}
\alias{check_aquifer}
\title{Check aquifer properties}
\usage{
check_aquifer(aquifer, standard_columns = NULL)
}
\description{
Check aquifer properties
}
\examples{
aquifer <- define_aquifer("confined",1e-4)
check_aquifer(aquifer)

aquifer2 <- define_aquifer("unconfined",1e-4)
check_aquifer(aquifer2,c("bounds","Ksat"))
check_aquifer(aquifer2,c("Ksat"))
}
\keyword{internal}
