% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_assess.R
\name{check_bounds}
\alias{check_bounds}
\title{Check boundaries}
\usage{
check_bounds(bounds)
}
\description{
Check that boundaries have slopes that are normal to each other
}
\examples{
check_bounds(tibble(m=c(1,1,-1,-1)))
check_bounds(tibble(m=c(1,0.9,-1,-1)))
}
\keyword{internal}
