% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_geoprocessing.R
\name{check_point_side_of_line}
\alias{check_point_side_of_line}
\title{Check point is in bounds}
\usage{
check_point_side_of_line(line, x, y)
}
\arguments{
\item{line}{list containing m and m}

\item{x}{x coordinates}

\item{y}{y coordinates}
}
\value{
Returns +1 for any point above the line, -1 for any point below the line, and 0 for a point on the line. If
the line is vertical (m=Inf), returns +1 for points to the right of the line and -1 for points to the left of the line.
}
\description{
Check point is in bounds
}
\examples{
\dontrun{
x <- 100
y <- 100
line <- list(m=Inf,b=2)
check_point_side_of_line(list(m=Inf,b=2),0,0)
check_point_side_of_line(list(m=Inf,b=2),2,0)
check_point_side_of_line(list(m=Inf,b=2),3,0)
check_point_side_of_line(list(m=1,b=2),0,0)
check_point_side_of_line(list(m=1,b=2),0,2)
check_point_side_of_line(list(m=1,b=2),0,3)
}
}
\keyword{internal}
