% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_preparation.R
\name{define_aquifer}
\alias{define_aquifer}
\title{Define aquifer}
\usage{
define_aquifer(aquifer_type, Ksat, ...)
}
\arguments{
\item{aquifer_type}{"confined" or "unconfined"}

\item{Ksat}{Saturated hydraulic conductivity}
}
\value{
This function returns an S3 "aquifer" object that behaves as a list and contains
 the named items:
\itemize{
 \item aquifer_type: "confined" or "unconfined"
 \item Ksat: Saturated hydraulic conductivity
 \item h0: Hydraulic head of the undisturbed aquifer (i.e., no pumping)
 \item z0: Thickness of the aquifer (applies only to confined aquifers)
 \item bounds: Boundaries of the aquifer, if any (generated using \code{define_boundaries()})
}
}
\description{
Define an aquifer as confined or unconfined and with saturated hydraulic conductivity.
Other optional parameters can be defined including boundaries and recharge.
}
\examples{
a <- define_aquifer("confined",1e-4)
}
