% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_preparation.R
\name{define_wells}
\alias{define_wells}
\title{Define wells}
\usage{
define_wells(wells_df = NULL, ...)
}
\arguments{
\item{wells_df}{A data.frame (or tibble, or sf with point geometry)
containing properties of wells. If wells_df is an sf object, the
coordinates for each well are extracted and the sf geometry is removed.}

\item{...}{Named vector inputs to be added to the wells object. Names that
replicate columns in \code{wells_df} supplant the columns in wells_df,
except \code{wID} and \code{well_type} which are generated automatically.
The length of inputs should generally be the same as the number of wells.
However, if an input is repeated for all wells, it can be input with length 1.}
}
\value{
This function returns a data.frame containing the columns:
\itemize{
 \item wID: Well ID
 \item Q: Pumping rate
 \item R: Radius of influence
 \item diam: Diameter of the well
 \item x, y: Cartesian coordinates
 \item well_type: Pumping (Q < 0), Injection (Q > 0), or Non-operational (Q = 0) well
}
These columns are followed by any other columns desired for the wells.
}
\description{
Define wells in the aquifer with required parameters.
}
\examples{
define_wells(Q=rep(0,3))

wells_df <- data.frame(x=1:3,y=1:3,Q=11:13,diam=0.5,id=1001:1003)
wells <- define_wells(wells_df)
print(wells) # prints a tibble of the wells

define_wells(wells_df,Note="note",Name=c("N1","N2","N3"))
define_wells(wells_df,R=1e4,Q=-21:-23) # include R and replace Q in wells_df
}
