% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_geoprocessing.R
\name{gen_circles}
\alias{gen_circles}
\title{Generate an object that acts like a circle}
\usage{
gen_circles(df)
}
\arguments{
\item{df}{data.frame with columns x, y, r}
}
\description{
Generate an object that acts like a circle
}
\examples{
df <- data.frame(x=1:3,y=1:3,roi=c(0.5,1,1.5))
circles <- gen_circles(df \%>\% rename(r=roi))
ggplot(circles) + geom_polygon(aes(x,y,group=id),color="black",alpha=0.5) + coord_equal()
}
