% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_imaging.R
\name{gen_well_image_type}
\alias{gen_well_image_type}
\title{Get pumping sign}
\usage{
gen_well_image_type(N, well_image, bound_type1, bound_type2,
  type = "text")
}
\arguments{
\item{N}{Vector of levels, N}

\item{well_image}{Character vector of well_image ("Actual", "Image (+Q)", or "Image (-Q)")}

\item{bound_type1}{Character vector of bound_type for first mirror boundary ("CH" or "NF")}

\item{bound_type2}{Character vector of bound_type for second mirror boundary ("CH" or "NF")}
}
\description{
Get sign of pumping of image well relative to pumping of Actual well
(associated by orig_wID). Sign assumes the point is mirrored across parallel
boundaries of constant head and/or no-flow.
}
\examples{
gen_well_image_type(N=rep(1:2,2),well_image=rep("Actual",4),
  bound_type1=c("NF","NF","CH","CH"),bound_type2=c("CH","CH","NF","NF"))
well_image <- "Actual"
bound_type1 <- "NF"
bound_type2 <- "NF"
gen_well_image_type(N,well_image,bound_type1,bound_type2)
gen_well_image_type(rep(1,4),"Actual","NF","NF")
gen_well_image_type(1:8,"Actual","NF","NF")
gen_well_image_type(1:8,"Actual","CH","NF")
gen_well_image_type(1:8,"Actual","NF","CH")
gen_well_image_type(1:8,"Image (+Q)","NF","CH")
gen_well_image_type(1:8,"Image (-Q)","NF","CH")
gen_well_image_type(1:8,"Actual","CH","CH")
}
