% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_particle_tracking.R
\name{get_capture_zone}
\alias{get_capture_zone}
\title{Get capture zone}
\usage{
get_capture_zone(wells, aquifer, wIDs, t_max, n_particles = 8,
  type = "all")
}
\arguments{
\item{wells}{Wells \code{data.frame} object, containing well images}

\item{aquifer}{Aquifer as an \code{aquifer} object, with \code{Ksat}, porosity, \code{n}, and boundaries (which are used to calculate gridded velocities)}

\item{t_max}{number of days to run particle tracking}

\item{n_particles}{number of particles to initialize for each well}

\item{type}{specifie which results to return. One of \code{"all"}, \code{"paths"}, \code{"endpoints"}, or \code{"smoothed"}. See details.}

\item{wID}{numeric value (or vector) containing wID for wells at which to generate capture zones.}
}
\description{
Get the capture zone for one or more wells
}
\details{
\itemize{
  \item "all": Returns a list containing all three named objects described below.
  \item "paths": Returns a data.frame of particle paths
  \item "endpoints": Returns the end location of each particle after tracking
  \item "smoothed": Returns a data.frame with smoothed path intersecting the \code{endpoints}
}
}
\examples{
aquifer <- aquifer_example
wells <- define_wells(wells_example[c(3,4,5),]) \%>\% generate_image_wells(aquifer)
get_capture_zone(wells,aquifer, wIDs = c(1,2))
ggplot() + geom_point(data=wells,aes(x,y,color=wID))
}
