% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_particle_tracking.R
\name{get_capture_zone_regions}
\alias{get_capture_zone_regions}
\title{Get capture zone regions}
\usage{
get_capture_zone_regions(particle_paths, aquifer, ...)
}
\arguments{
\item{type}{specifies which results to return. One of \code{"all"}, \code{"paths"}, \code{"endpoints"}, or \code{"smoothed"}. See details.}
}
\description{
Get well capture zone regions
}
\details{
\itemize{
  \item "all": Returns a list containing all three named objects described below.
  \item "paths": Returns a data.frame of particle paths
  \item "endpoints": Returns the end location of each particle after tracking
  \item "smoothed": Returns a data.frame with smoothed path intersecting the \code{endpoints}
}
}
