% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_particle_tracking.R
\name{get_confined_flowlines}
\alias{get_confined_flowlines}
\title{Get flowlines in confined aquifer}
\usage{
get_confined_flowlines(wells, aquifer, nominal_levels = 40,
  flow_dim = c(100, 100))
}
\arguments{
\item{wells}{Wells object}

\item{aquifer}{Aquifer object}

\item{flow_dim}{Dimensions for generating raster for generating flowlines}
}
\description{
Get flowlines in confined aquifer
}
\examples{
# define aquifer
bounds_df <- data.frame(bound_type=c("NF","NF","NF","NF"),m=c(Inf,0,Inf,0),b=c(0,1000,1000,0))
aquifer <- define_aquifer("confined",1e-3,bounds=bounds_df,h0=100)

# define wells and well images
wells_df <- data.frame(x=c(300,700),y=c(200,600),diam=1) \%>\%
  dplyr::mutate(R=1000,Q=-1/dplyr::n())
wells <- define_wells(wells_df) \%>\% generate_image_wells(aquifer)

ggplot() +
  geom_segment(data=aquifer_unconfined$bounds,aes(x1,y1,xend=x2,yend=y2,color=bound_type)) +
  geom_point(data=wells \%>\% filter(wID==orig_wID),aes(x,y),shape=21) +
  geom_path(data=cl0,aes(x,y,color=as.factor(line))) +
  coord_equal()
}
