% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_geoprocessing.R
\name{get_edges_from_vertices}
\alias{get_edges_from_vertices}
\title{Vertices to edges}
\usage{
get_edges_from_vertices(vertices)
}
\arguments{
\item{vertices}{A data.frame of x, y, and bID}
}
\value{
This function returns bounds object from x, y vertices, with bID 1:4
}
\description{
Convert vertices to edges. It prioritizes quadrangles -- it will close the polygon if
there are 4 vertices.
}
\examples{
vertices <- data.frame(x=c(0,1),y=c(0,1),bID=1:2)
get_edges_from_vertices(vertices)

vertices <- data.frame(x=c(0,1,1),y=c(0,1,0.5),bID=1:3)
get_edges_from_vertices(vertices)

vertices <- data.frame(x=c(0,0,1,1),y=c(0,1,1,0),bID=1:4)
get_edges_from_vertices(vertices)
}
