% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_potentials.R
\name{get_flowdir_raw}
\alias{get_flowdir_raw}
\title{Get flow direction in confined aquifer}
\usage{
get_flowdir_raw(loc, wells, aquifer)
}
\arguments{
\item{loc}{coordinates \code{data.frame} with columns labeled 'x' and 'y', or as vector as c(x,y), with units of [m]}

\item{wells}{wells object with each row containing rate Q [m^3/s], diam [m],
radius of influence R [m], & coordinates x [m], y [m]}

\item{aquifer}{Afuifer object containing aquifer_type, h0, Ksat, bounds, z0 (for confined case only)}
}
\value{
The output is the cumulative effect at \code{loc} of all \code{wells} on the flow
  Note: if the \code{loc} is contained within the diameter of a well, the distance between the location
  and that well is automatically adjusted to the edge of the well screen (i.e., well$diam/2). Similar any well-location
  distance that exceeds the radius of influence of the well, R, is set equal to R
}
\description{
Get the cumulative effect of all wells on flow at a single location, output as dh/dx and dh/dy.
}
\examples{
wells <- define_wells(x=c(0,0.5),y=c(0,0.25),Q=c(1e-3,-2e-3),diam=c(0.75,0.8),R=c(300,300))
aquifer <- define_aquifer(h0=0,Ksat=0.00001,z0=30,aquifer_type="confined")
get_flowdir(loc=c(5,5),wells,aquifer)
get_flowdir_raw(loc=c(5,5),wells,aquifer)
get_flowdir_raw(loc=c(5,5),NULL,aquifer)

grid_pts <- expand.grid(x=seq(0,10,by=5),y=seq(0,10,by=5))
get_flowdir(loc=grid_pts,wells,aquifer)
get_flowdir_raw(loc=grid_pts,wells,aquifer)
get_flowdir_raw(loc=grid_pts,NULL,aquifer)

recharge_params <- list(recharge_type="F",recharge_vector=c(0,0,-1,-1),flow=1e-3,x0=0,y0=0)
aquifer_norecharge <- define_aquifer("confined",1e-3,h0=0,z0=1)
aquifer_recharge <- define_aquifer("confined",1e-3,h0=0,z0=1,recharge=recharge_params)
(a <- get_flowdir(c(-1,-1),wells,aquifer_recharge))
(b <- get_flowdir(c(-1,-1),wells,aquifer_norecharge))
(d <- get_flowdir(c(-1,-1),NULL,aquifer_recharge))
b + d

recharge_params <- list(recharge_type="D",recharge_vector=c(0,0,-1,-1),flow_main=1,flow_opp=1,x0=0,y0=0)
aquifer_norecharge <- define_aquifer("confined",1,h0=0,z0=1)
aquifer_recharge <- define_aquifer("confined",1,h0=0,z0=1,recharge=recharge_params)
loc <- expand.grid(x=-1:1,y=-1:1)
(a <- get_flowdir(loc,wells,aquifer_recharge))
(b <- get_flowdir(loc,wells,aquifer_norecharge))
(d <- get_flowdir(loc,NULL,aquifer_recharge))
b + d - a

recharge_params <- list(recharge_type="D",recharge_vector=c(10,10,11,10),flow_main=1,flow_opp=1,x0=0,y0=0)
aquifer_norecharge <- define_aquifer("confined",1,h0=0,z0=1)
aquifer_recharge <- define_aquifer("confined",1,h0=0,z0=1,recharge=recharge_params)
loc <- expand.grid(x=9:11,y=9:11)
(d <- get_flowdir(loc,NULL,aquifer_recharge))
cbind(loc,d)
}
