% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_geoprocessing.R
\name{get_intersection}
\alias{get_intersection}
\title{Get intersection of two lines specified by slope m and intercept, b}
\usage{
get_intersection(m1, b1, m2, b2)
}
\description{
Get intersection of two lines specified by slope m and intercept, b
}
