% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anema_imaging.R
\name{get_mirror_point}
\alias{get_mirror_point}
\title{Mirror a point across a boundary}
\usage{
get_mirror_point(point, boundary, new_pID = NA)
}
\arguments{
\item{point}{single point as a list (or sf object single point feature)
containing coordinates (point$x0, point$y0), and pumping rate (point$rate),
which is positive for injection, negative for pumping. Should also have:
diam, path, origin.}

\item{boundary}{single line as a list (or sf object single line feature)
containing slope (as boundary$m) and intercept (as boundary$b), as well as
the boundary type (bound_type) as "CH" for constant head or "NF" for no flow}

\item{new_pID}{integer ID for the new point to be created}
}
\value{
a row (same class as point input) containing the new imaged point and
  the following columns: x0, y0, rate, diam, path, origin, transform
  (boundary type),
  source_bound
}
\description{
Mirror a point across a boundary, specified as constant head (CH) or no flow
(NF) and defined by the line \eqn{m x + b})
}
\examples{
point <- data.frame(pID=1,x0=0,y0=0,rate=0.5,diam=0.75,path=1,origin=1)
boundary <- data.frame(bID=1,m=-1,b=1,bound_type="NF")
get_mirror_point(point, boundary, new_pID=2)
}
