% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_geoprocessing.R
\name{get_nearest_point_on_line}
\alias{get_nearest_point_on_line}
\title{Get nearest point on line}
\usage{
get_nearest_point_on_line(loc, m, b)
}
\arguments{
\item{loc}{Points as \code{c(x,y)} or a \code{data.frame} with columns \code{x} and \code{y}}

\item{m}{Slope of the line}

\item{b}{Intercept of the line (if \code{m} is \code{Inf}, \code{b} should be the x-intercept)}
}
\description{
Get nearest point on
}
\examples{
get_nearest_point_on_boundary(c(1,1),)
bounds <- data.frame(bound_type=c("CH","NF","NF","NF"),m=c(Inf,0,Inf,0),b=c(0,0,100,100)) \%>\% define_bounds()
loc <- c(150,150)
get_nearest_point_on_line(loc,bounds[1,]$m,bounds[1,]$b)

bounds <- data.frame(bound_type=c("CH","NF","NF","NF"),m=c(-2,0.5,-2,0.5),b=c(0,0,100,20)) \%>\% define_bounds()
loc <- data.frame(x=c(-100,0,100),y=c(-150,0,150))
nearest <- get_nearest_point_on_line(loc,bounds[1,]$m,bounds[1,]$b)
ggplot() +
   geom_segment(data=bounds,aes(x1,y1,xend=x2,yend=y2))  +
   geom_abline(data=bounds[1,],aes(slope=m,intercept=b),linetype="dashed")  +
   geom_point(data=loc,aes(x,y),color="black")  +
   geom_point(data=nearest,aes(x,y),color="red",alpha=0.5) +
   coord_equal()

loc_rep <- data.frame(x=rep(0,4),y=rep(75,4))
nearest <- get_nearest_point_on_line(loc_rep,m=bounds$m,b=bounds$b)
ggplot() +
   geom_segment(data=bounds,aes(x1,y1,xend=x2,yend=y2))  +
   geom_abline(data=bounds,aes(slope=m,intercept=b),linetype="dashed")  +
   geom_point(data=loc_rep,aes(x,y),color="black")  +
   geom_point(data=nearest,aes(x,y),color="red",alpha=0.5) +
   coord_equal()
}
\keyword{internal}
