% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_geoprocessing.R
\name{get_perpendicular_line}
\alias{get_perpendicular_line}
\title{Get perpendicular line}
\usage{
get_perpendicular_line(m, x, y)
}
\arguments{
\item{m}{Slope of line}

\item{x}{Coordinate through which perpendicular line should pass}

\item{y}{Coordinate through which perpendicular line should pass}
}
\value{
Returns a list containing $m and $b, the slope and intercept of a perpendicular
line that passes through x, y.
}
\description{
Get perpendicular in from slope and point
}
\examples{
get_perpendicular_line(Inf,2,3)
get_perpendicular_line(0,2,3)
get_perpendicular_line(1/2,2,2)
get_perpendicular_line(-2,2,2)
}
\keyword{internal}
