% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_geoprocessing.R
\name{get_point_on_quandrangle}
\alias{get_point_on_quandrangle}
\title{Get point on quadrangle}
\usage{
get_point_on_quandrangle(x, y, axis)
}
\description{
The complete quadrangle is comprised of 6 vertices (intersection points)
of the edges of the quadrangle. But only 4 of these are the vertices of
the quadrangle. See http://mathworld.wolfram.com/CompleteQuadrilateral.html.
If a line has 3 intersection points, the middle one is definitely a vertex
of the quadrangle. If less than 3, then both are intersection points.
Choose the second one.
}
\keyword{internal}
