% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_potentials.R
\name{get_potential}
\alias{get_potential}
\title{Get cumulative effect of wells at location}
\usage{
get_potential(loc, wells, Ksat, z0, aquifer_type)
}
\arguments{
\item{loc}{coordinates \code{data.frame} with columns labeled 'x' and 'y', or as vector as c(x,y), with units of [m]}

\item{wells}{wells object with each row containing rate Q [m^3/s], diam [m],
radius of influence roi [m], & coordinates x0 [m], y0 [m]}

\item{Ksat}{saturated hydraulic conductivity [m/s]}

\item{z0}{thickness of the confined aquifer (confining case only)}

\item{aquifer_type}{"confined" or "unconfined". Determines the calculation
and output units.}
}
\value{
The output is the cumulative effect at \code{loc} of all \code{wells} on the hydraulic head [units=m] if
  \code{aquifer_type="confined"} or discharge potential [m^2] if
  \code{aquifer_type="unconfined"}.

  Note: if the \code{loc} is contained within the diameter of a well, the distance between the location
  and that well is automatically adjusted to the edge of the well screen (i.e., well$diam/2). Similar any well-location
  distance that exceeds the radius of influence of the well, R, is set equal to R
}
\description{
Get the cumulative effect of all wells at a singled location, output as head (confined aquifer) or discharge potential (unconfined aquifer).
}
\examples{
# Single test location
wells <- data.frame(x0=c(0,0.5),y0=c(0,0.25),rate=c(1e-4,-2e-4),diam=c(0.75,0.8),roi=c(300,300))
get_potential(wells,loc=c(50,50),Ksat=0.00001,z0=10,aquifer_type="confined")
get_potential(wells,loc=c(50,50),Ksat=0.00001,aquifer_type="unconfined")

# Multiple test locations
wells <- data.frame(x0=c(-10,10),y0=c(-10,10),rate=c(1e-3,-1e-3),diam=c(0.1,0.1),roi=c(300,300))
grid_pts <- data.frame(x=c(-11,0,11),y=c(-11,0,11))
get_potential(grid_pts,wells,Ksat=0.00001,aquifer_type="unconfined")
}
