% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_potentials.R
\name{get_potential_differential}
\alias{get_potential_differential}
\title{Get cumulative effect of wells at location}
\usage{
get_potential_differential(loc, wells, aquifer)
}
\arguments{
\item{loc}{coordinates \code{data.frame} with columns labeled 'x' and 'y', or as vector as c(x,y), with units of [m]}

\item{wells}{wells object with each row containing rate Q [m^3/s], diam [m],
radius of influence R [m], & coordinates x [m], y [m]}

\item{aquifer}{Afuifer object containing aquifer_type, h0, Ksat, bounds, z0 (for confined case only)}
}
\value{
The output is the cumulative effect at \code{loc} of all \code{wells} on the hydraulic head [units=m] if
  \code{aquifer_type="confined"} or discharge potential [m^2] if
  \code{aquifer_type="unconfined"}.

  Note: if the \code{loc} is contained within the diameter of a well, the distance between the location
  and that well is automatically adjusted to the edge of the well screen (i.e., well$diam/2). Similar any well-location
  distance that exceeds the radius of influence of the well, R, is set equal to R
}
\description{
Get the cumulative effect of all wells at a singled location, output as head (confined aquifer) or discharge potential (unconfined aquifer).
}
\examples{
# Single test location
wells <- define_wells(x=c(0,0.5),y=c(0,0.25),Q=c(1e-4,-2e-4),diam=c(0.75,0.8),R=c(300,300))
aquifer <- define_aquifer(aquifer_type="confined",Ksat=0.00001,h0=0,z0=30)

wells <- rbind(well1,well2)
get_potential_differential(loc=c(50,50),well1,aquifer)
get_potential_differential(loc=c(50,50),well2,aquifer)

# Multiple test locations
wells <- data.frame(x=c(-10,10),y=c(-10,10),Q=c(1e-3,-1e-3),diam=c(0.1,0.1),R=c(300,300))
grid_pts <- data.frame(x=c(-11,0,11),y=c(-11,0,11))
aquifer_unconfined <- define_aquifer(aquifer_type="confined",Ksat=0.00001,h0=0,z0=30)
get_potential_differential(grid_pts,wells,aquifer_unconfined)

get_potential_differential(c(1,1))
get_potential_differential(data.frame(x=c(1,1),y=c(2,2)))
}
