% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_imaging.R
\name{get_pumping_sign}
\alias{get_pumping_sign}
\title{Get pumping sign}
\usage{
get_pumping_sign(well_image)
}
\description{
Get pumping sign
}
\examples{
get_pumping_sign("Actual")
get_pumping_sign("Image (+Q)")
get_pumping_sign("Image (-Q)")
}
