% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_geoprocessing.R
\name{get_quad_vertices}
\alias{get_quad_vertices}
\title{Get vertices of a quadrangle}
\usage{
get_quad_vertices(boundaries)
}
\arguments{
\item{boundaries}{a data.frame containing 4 lines (rows) defined, respectively, by x1, y1, x2, y2}
}
\value{
Returns a data.frame vertices for each of the 4 boundaries. The boundaries
are given an id, and there is one row in the output which identifies the opposide bound
(with which there is no intersection).
}
\description{
Get the 4 vertices correspondingt to quadrangle of 4 lines
}
\examples{
boundaries <- data.frame(x1=c(0,10,13,1),y1=c(0,10,9,-1),x2=c(10,13,1,0),y2=c(10,9,-1,0))
quad_vertices <- get_quad_vertices(boundaries)
ggplot() + geom_segment(data=boundaries,aes(x1,y1,xend=x2,yend=y2)) +
  geom_point(data=quad_vertices,aes(x,y,shape=as.factor(bound_id),color=as.factor(bound_id)))
}
