% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_geoprocessing.R
\name{get_quad_vertices_wide}
\alias{get_quad_vertices_wide}
\title{Get quadrangle vertices}
\usage{
get_quad_vertices_wide(bounds)
}
\description{
Get quadrangle vertices in wide format
}
\keyword{internal}
