% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_potentials.R
\name{get_radius_of_influence}
\alias{get_radius_of_influence}
\title{Estimate the radius of influence of a well}
\usage{
get_radius_of_influence(..., method)
}
\arguments{
\item{...}{Variable parameters, depending on the method used}

\item{method}{String containing the name of the desired method}
}
\value{
A numeric value indicating the horizontal radius of influence of the
  well.
}
\description{
Estimate the radius of influence of a well
}
\section{Methods}{
 The methods below are taken from Fileccia (2015). Acque
  Sotterranee - Italian Journal of Groundwater. DOI 10.7343/AS-117-15-0144.
  The following strings can be input for the \code{method} variable, and must be accompanied
  by the corresponding variables as part of the \code{...} input:

  \describe{
    \item{"cooper-jacob":}{\eqn{R=\sqrt{2.25 Tr t / S}}, for confined aquifer after short pumping period. (Cooper and Jacob, 1946)}
    \item{"aravin-numerov":}{\eqn{R=\sqrt{1.9 Ksat h t / n}}, for unconfined aquifers (Aravin and Numerov, 1953)}
    \item{"sichardt":}{\eqn{R=3000 s \sqrt{Ksat}}, Sichardt formula for unconfined aquifers (Cashman and Preene, 2001)}
  }

  Where:
  \itemize{
    \item R = radius of influence [m]
    \item Tr = transmissivity [m^2/s]
    \item t = time [s]
    \item S = storage
    \item h = height of the water table above substratum [m]
    \item n = effective porosity
    \item Ksat = saturated hydraulic conductivity [m/s]
    \item s = drawdown in the borehole [m]
  }
}

\examples{
get_radius_of_influence(Tr=0.01,t=3600*12,S=1,method="cooper-jacob")
get_radius_of_influence(Ksat=0.0001,h=50,t=3600*12,n=0.4,method="aravin-numerov")
get_radius_of_influence(Ksat=0.0001,s=10,method="sichardt")
}
