% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_recharge.R
\name{get_recharge_flow_potential}
\alias{get_recharge_flow_potential}
\title{Recharge flow potential function}
\usage{
get_recharge_flow_potential(loc, aquifer)
}
\arguments{
\item{loc}{Location for evaluating Vector location as c(x,y) or data.frame containing x, y columns}

\item{Aquifer}{Aquifer containing containing \code{aquifer_type},\code{h0} and \code{recharge} (see Details)}
}
\value{
Returns hydraulic head (confined aquifers) or discharge potential (unconfined aquifers) at the location(s).
}
\description{
Recharge flow potential function
}
\details{
Recharge parameters must contain \code{x0}, \code{y0}, \code{x_term}, and \code{y_term}. For confined aquifers,
the last two are:
\itemize{
\item \code{x_term}: -flow * cos(theta) * sign(dx) / (Ksat * z0)
\item \code{y_term}: -flow * sin(theta) * sign(dy) / (Ksat * z0)
}
For unconfined aquifers,
these parameters are:
\itemize{
\item \code{x_term}: -2 * flow * cos(theta) * sign(dx) / Ksat
\item \code{y_term}: -2 * flow * sin(theta) * sign(dy) / Ksat
}
}
\examples{
## Flow - confined aquifer
recharge_params <- list(recharge_type="F",recharge_vector=c(0,0,1,2),flow=1,x0=0,y0=0)
aquifer <- define_aquifer("confined",1,h0=0,z0=1,recharge=recharge_params)
get_recharge_flow_potential(c(1/sqrt(2),1/sqrt(2)), aquifer)

loc <- expand.grid(x=-2:2,y=-2:2)
loc$h <- get_recharge_flow_potential(loc, aquifer)
library(ggplot2)
ggplot(loc) + geom_raster(aes(x,y,fill=h)) + scale_fill_gradient2()

recharge_params <- list(recharge_type="F",recharge_vector=c(-1,-5,0,0),flow=1,x0=0,y0=0)
aquifer <- define_aquifer("confined",1,h0=10,z0=1,recharge=recharge_params)
loc <- expand.grid(x=-100:100,y=-100:100)
loc$h <- get_recharge_flow_potential(loc, aquifer)
ggplot(loc) + geom_raster(aes(x,y,fill=h)) + scale_fill_gradient2()

recharge_params <- list(recharge_type="F",recharge_vector=c(-1,-5,0,0),flow=1,x0=0,y0=0)
aquifer <- define_aquifer("unconfined",1e-1,h0=50,recharge=recharge_params)
loc <- expand.grid(x=-100:100,y=-100:100)
loc$h2 <- get_recharge_flow_potential(loc, aquifer)
loc <- loc \%>\% dplyr::mutate(h=sqrt(h2))
ggplot(loc) + geom_raster(aes(x,y,fill=h)) + scale_fill_gradient2(midpoint=50)
}
