% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_recharge.R
\name{get_recharge_hydraulic_potential}
\alias{get_recharge_hydraulic_potential}
\title{Get hydraulic potential due to recharge}
\usage{
get_recharge_hydraulic_potential(loc, aquifer)
}
\arguments{
\item{loc}{Location for evaluating Vector location as c(x,y) or data.frame containing x, y columns}

\item{Aquifer}{Aquifer containing containing \code{h0} and \code{recharge$params} (see Details)}
}
\value{
Returns hydraulic potential at the location(s) using helper functions for each recharge_type / aquifer_type:
\itemize{
\item 1. Constant flow, "F": \code{get_recharge_flow_potential()}
\item 2. Head boundaries, "H": not yet implemented
\item 3. Recharge divide, "D": \code{get_recharge_divide_potential()}
}
}
\description{
Get hydraulic potential due to recharge
}
\details{
The aquifer and aquifer$recharge must be correctly specified using \code{define_aquifer} and/or \code{define_recharge}.
}
\examples{
## Flow - confined aquifer
recharge_params <- list(recharge_type="F",recharge_vector=c(0,0,1,1),flow=1,x0=0,y0=0)
aquifer <- define_aquifer("confined",1,h0=0,z0=1,recharge=recharge_params)
get_recharge_hydraulic_potential(c(1/sqrt(2),1/sqrt(2)), aquifer)

loc <- expand.grid(x=-2:2,y=-2:2)
loc$h <- get_recharge_hydraulic_potential(loc, aquifer)
library(ggplot2)
ggplot(loc) + geom_raster(aes(x,y,fill=h)) + scale_fill_gradient2()

recharge_params <- list(recharge_type="F",recharge_vector=c(-1,-5,0,0),flow=1,x0=0,y0=0)
aquifer <- define_aquifer("confined",1,h0=10,z0=1,recharge=recharge_params)
loc <- expand.grid(x=-100:100,y=-100:100)
loc$h <- get_recharge_hydraulic_potential(loc, aquifer)
ggplot(loc) + geom_raster(aes(x,y,fill=h)) + scale_fill_gradient2()
}
