% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_geoprocessing.R
\name{get_rectangle}
\alias{get_rectangle}
\title{#' Get intersection of two lines
#'
#' Get the intersection of two lines described by m, b
get_intersection()}
\usage{
get_rectangle(boundaries)
}
\arguments{
\item{boundaries}{a data.frame containing 4 lines (rows) defined, respectively, by x1, y1, x2, y2}
}
\value{
Returns a data.frame containing slope and intercept for four edges of a rectangle. The
rectangle is determined by (1) identifying the quadrangle of the input \code{boundaries}, (2)
selecting the midpoints of each edge of the quadrangle, (3) determining the "long" axis of the
quadrangle, which becomes the long axis of the rectangle, (4) calculating the slope of the long
and short axes of the rectangle (at right angles), then generating lines with these slopes through
the midpoints.
}
\description{
Convert quadrangle to rectangle
}
\details{
Convert 4 lines to a rectangular box with only right angles
}
\examples{
boundaries <- data.frame(x1=c(0,10,13,1),y1=c(0,10,9,-1),x2=c(10,13,1,0),y2=c(10,9,-1,0))
rect_boundaries <- get_rectangle(boundaries) \%>\% mutate(id=as.factor(row_number()))
ggplot() + geom_segment(data=boundaries,aes(x1,y1,xend=x2,yend=y2)) +
  geom_abline(data=rect_boundaries,aes(slope=m,intercept=b,color=id))
}
