% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_potentials.R
\name{get_row_as_vector}
\alias{get_row_as_vector}
\title{Get row of a data.frame as a vector}
\usage{
get_row_as_vector(df, row = 1)
}
\description{
return the row of a data.frame or tibble as a vector (used for extracting x,y coordinates)
}
\keyword{internal}
