% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_assess.R
\name{get_segment_seq}
\alias{get_segment_seq}
\title{Get segment sequence}
\usage{
get_segment_seq(segment, length.out = 10)
}
\arguments{
\item{segment}{A list containing x1, y1, x2, y2, and sID}
}
\description{
Get segment sequence
}
\examples{
get_segment_seq(list(sID=1,x1=0,y1=0,x2=1,y2=1),length.out=5)
}
\keyword{internal}
