% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_wrappers.R
\name{get_single_drawdown_relationship}
\alias{get_single_drawdown_relationship}
\title{Get drawdown relationship}
\usage{
get_single_drawdown_relationship(wells, aquifer, group_column,
  weights_column, loc_group, pump_group)
}
\arguments{
\item{wells}{Wells data.frame or tibble}

\item{aquifer}{Aquifer object}

\item{group_column}{Name of column to identify groups of wells (no quotes)}

\item{weights_column}{Name of column to identify weights of well pumping (no
quotes). Required -- create column equal to 1 if equal weights desired.}

\item{loc_group}{Value in wells$group_column where drawdown should be measured}

\item{pump_group}{Value in wells$group_column which identifies wells}
}
\value{
Returns a list with two values:
pot, which is the potential differential due to unit pumping from the pumping wells in pump_group.

var, which is given as D_i_j, or PHI_i_j -- the average drawdown at wells i due to wells j.

The unit pumping at the pumping wells is weighted by values in weights_column for the pump_group.
The average drawdown is also weighted by values in weights_column, for the loc_group.
}
\description{
Get drawdown relationship at locations locations due to unit pumping from wells. Include weights on location and wells.
}
\examples{
# define aquifer
bounds_df <- data.frame(bound_type=c("CH","NF","NF","NF"),m=c(Inf,0,Inf,0),b=c(0,1000,1000,0))
aquifer_confined <- define_aquifer("unconfined",1e-3,bounds=bounds_df,h0=100)

# define wells and well images
set.seed(30)
wells_df <- data.frame(x=runif(8,0,1000),y=runif(8,0,1000),diam=1) \%>\%
  mutate(R=1000,  # t = 1 year
         country=factor(y>500,levels=c(F,T),labels=c("A","B"))) \%>\%
  group_by(country) \%>\%
  mutate(weights=1,Q=1/n()) \%>\% group_by()
wells <- define_wells(wells_df) \%>\% generate_image_wells(aquifer_confined)

get_single_drawdown_relationship(wells, aquifer_confined, group_column=country, weights_column=weights,loc_group="S", pump_group="F")

ggplot() +
  geom_segment(data=aquifer_confined$bounds,aes(x1,y1,xend=x2,yend=y2,color=bound_type)) +
  geom_abline(slope=0,intercept=500,linetype="dashed") +
  geom_point(data=wells \%>\% filter(wID==orig_wID),aes(x,y,fill=country),shape=21) +
  coord_equal()
}
