% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_geoprocessing.R
\name{get_slope_intercept}
\alias{get_slope_intercept}
\title{Get slope and intercept}
\usage{
get_slope_intercept(x1, y1, x2 = NULL, y2 = NULL, m = NULL)
}
\description{
Get slope and intercept of a line defined by x1, y1, x2, y2
}
\examples{
\dontrun{
get_slope_intercept(0,0,1,2)
get_slope_intercept(0,-0,2,-1)
get_slope_intercept(0,0,0,3)
get_slope_intercept(0,2,3,2)
get_slope_intercept(1,1,m=-1)
get_slope_intercept(1,1,m=Inf)
}
}
\keyword{internal}
