% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_particle_tracking.R
\name{get_theta}
\alias{get_theta}
\title{Get angle from x, y}
\usage{
get_theta(x, y)
}
\description{
Get angle from x,y
}
\examples{
\dontrun{
get_theta(-1,0)
get_theta(1,0)
get_theta(-1,1)/pi
get_theta(0,1)/pi
get_theta(c(1,1,0,-1,-1),c(0,1,1,1,0))/pi
}
}
\keyword{internal}
