% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_potentials.R
\name{get_well_effect}
\alias{get_well_effect}
\title{Get the effect of a single well}
\usage{
get_well_effect(well, loc, Ksat, z0 = NULL, aquifer_type)
}
\arguments{
\item{well}{single well object containing rate Q [m^3/s], diam [m],
radius of influence roi [m], & coordinates x0 [m], y0 [m]}

\item{loc}{coordinates vector as c(x,y), with units of [m]}

\item{Ksat}{saturated hydraulic conductivity [m/s]}

\item{z0}{thickness of the confined aquifer (confining case only)}

\item{aquifer_type}{"confined" or "unconfined". Determines the calculation
and output units.}
}
\value{
The output is the effect at \code{loc} of the \code{well} on the hydraulic head (in units of [m]) if
  \code{aquifer_type="confined"} or discharge potential [m^2] if
  \code{aquifer_type="unconfined"}.

  Note: if the \code{loc} is contained within the diameter of the well, the distance between the tested location
  and the well is automatically adjusted to the edge of the well screen (i.e., well$diam/2).
}
\description{
Get the effect of a single well on hydraulic potential or discharge
potential.
}
\examples{
well <- data.frame(pID=1,x0=0,y0=0,rate=1e-4,diam=0.75,roi=300)
get_well_effect(well,loc=c(50,50),Ksat=0.00001,z0=10,aquifer_type="confined")
get_well_effect(well,loc=c(50,50),Ksat=0.00001,aquifer_type="unconfined")
}
