% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_data.R
\docType{data}
\name{groundwater_district}
\alias{groundwater_district}
\title{Groundwater district anem-app scenario}
\format{a list object containing named items associated with anem-app}
\usage{
groundwater_district
}
\description{
anem-app scenario with four pumping wells near a river
}
\keyword{datasets}
