% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_data.R
\docType{data}
\name{municipal_contamination}
\alias{municipal_contamination}
\title{Municipal contamination anem-app scenario}
\format{a list object containing named items associated with anem-app}
\usage{
municipal_contamination
}
\description{
anem-app scenario with urban contamination at a well
}
\keyword{internal}
