% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_assess.R
\name{plot_bounds_behavior}
\alias{plot_bounds_behavior}
\title{Plot behavior on boundaries}
\usage{
plot_bounds_behavior(well_images, aquifer, length.out = 100)
}
\arguments{
\item{aquifer}{Afuifer object containing aquifer_type, h0, Ksat, bounds, z0 (for confined case only)}

\item{length.out}{The number of points to evaluate on each boundary}
}
\value{
Two ggplot objects that show behavior on the boundaries, one for head and one for flow
}
\description{
Plot behavior on boundaries with and without well images
}
\examples{
library(tidyverse)
aquifer <- aquifer_confined_example
well_images <- define_wells(wells_example) \%>\% generate_image_wells(aquifer)
p_domain <- ggplot() +
  geom_point(data=well_images,aes(x,y,fill=Q),color="black",size=2,shape=21) +
  scale_fill_gradient2(low="blue",high="red",mid="gray") +
  geom_segment(data=aquifer$bounds,aes(x1,y1,xend=x2,yend=y2,linetype=bound_type)) +
  coord_equal()
gg_list <- plot_bounds_behavior(well_images,aquifer,length.out=20)
gridExtra::grid.arrange(p_domain,gg_list$p_h,gg_list$p_f,nrow=1)
gg_list$table
gg_list$bounds_behavior
}
