% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_geoprocessing.R
\name{prep_bounds}
\alias{prep_bounds}
\title{Prepare boundaries with slope and intercept}
\usage{
prep_bounds(boundaries, get_rectangular = FALSE)
}
\arguments{
\item{boundaries}{An sf object of straight lines (single segments), or a
data.frame with column names x1, y1, x2, y2, representing the endpoints}
}
\value{
A data.frame containing slope (m) and intercept (b) for each line,
  along with original columns. If \code{boundaries} is a data.frame, the columns
  x1, y1, x2, y2 are automatically calculated using \code{prep_wells_sf} and
  added to the sf object before getting the slope and intercept. The function requires
  that the wells are labeled with a column of identifiers, pID. If it is not
  present, the function generates them.
}
\description{
Get slope (m) and intercept (b) of all line objects, and add boundary ID (bID)
}
\examples{
boundaries <- data.frame(x1=c(0,10,13,1),y1=c(0,10,9,-1),x2=c(10,13,1,0),y2=c(10,9,-1,0))
ggplot(boundaries) + geom_segment(aes(x1,y1,xend=x2,yend=y2))
}
