% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anema_geoprocessing.R
\name{prep_wells_sf}
\alias{prep_wells_sf}
\title{Convert wells sf object to dataframe}
\usage{
prep_wells_sf(wells_sf)
}
\arguments{
\item{wells}{wells sf object, with point features}

\item{results}{outputs a data.frame of wells with all of the attributes and coordinate as x0, y0}
}
\description{
Convert wells sf object to dataframe, with appropriate column coordinates
}
