% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_printing.R
\name{print.aquifer}
\alias{print.aquifer}
\title{S3 Print method for class "aquifer"}
\usage{
\method{print}{aquifer}(x, ...)
}
\arguments{
\item{x}{Object of type aquifer}

\item{...}{Additional, optional arguments to print S3 generic}
}
\description{
S3 Print method for class "aquifer"
}
