% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_geoprocessing.R
\name{print_data_frame_for_entry}
\alias{print_data_frame_for_entry}
\title{Print data frame for re-entry}
\usage{
print_data_frame_for_entry(df)
}
\description{
Print data frame for re-entry
}
\examples{
df <- data.frame(a=1:4,b=11:14,c=c("a","b","c","q"))
print_data_frame_for_entry(df)
}
