% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_shiny_related.R
\name{run_shiny}
\alias{run_shiny}
\title{Run anem shiny app}
\usage{
run_shiny(shiny_app, display.mode = "auto")
}
\description{
Run shiny app within package anem
}
\examples{
run_shiny("anem_shiny")
}
