% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_imaging.R
\name{set_pumping_sign}
\alias{set_pumping_sign}
\title{Get pumping sign}
\usage{
set_pumping_sign(well_image, bound_type, type = "text")
}
\description{
Get sign of pumping relative to pumping of Actual well (associated by orig_wID)
}
\examples{
set_pumping_sign("Actual","NF")
set_pumping_sign("Actual","CH")
set_pumping_sign("Image (+Q)","NF")
set_pumping_sign("Image (+Q)","CH")
set_pumping_sign("Image (-Q)","NF")
set_pumping_sign("Image (-Q)","CH")
}
\keyword{internal}
