% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_particle_tracking.R
\name{track_particle}
\alias{track_particle}
\title{Track particle in aquifer}
\usage{
track_particle(loc, wells, aquifer)
}
\arguments{
\item{loc}{Coordinate vector as \code{c(x, y)}}

\item{wells}{Wells \code{data.frame} containing well images}

\item{aquifer}{Aquifer as an \code{aquifer} object, with \code{Ksat} and porosity, \code{n}
Returns a data.frame containing the time and locations}
}
\description{
Track a particle in the aquifer from an original location to a pumping well,
aquifer boundary, or outside of the wells ROI
}
\examples{
bounds_df <- data.frame(bound_type=c("NF","NF","CH","NF"),m=c(Inf,0,Inf,0),b=c(0,1000,1000,0))
aquifer <- define_aquifer(aquifer_type="confined",Ksat=0.001,n=0.4,h0=0,z0=20,bounds=bounds_df)
wells <- data.frame(x=c(500,100),y=c(300,600),Q=c(-1e-1,-1e-1),diam=c(1,1),R=c(1000,100)) \%>\%
  define_wells() \%>\% generate_image_wells(aquifer)
gridded <- get_gridded_hydrodynamics(wells,aquifer,c(100,100),c(10,10))
system.time(particle_path <- track_particle(c(750,200), wells, aquifer))
particle_path <- track_particle(c(500,250), wells, aquifer)
ggplot() +
  geom_raster(data=gridded$head,aes(x,y,fill=head_m)) +
  geom_segment(data=aquifer$bounds,aes(x1,y1,xend=x2,yend=y2,color=bound_type)) +
  geom_point(data=wells \%>\% filter(wID==orig_wID),aes(x,y),shape=21) +
  geom_path(data=particle_path,aes(x,y),color="red") +
  coord_equal()
}
