% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_wrappers.R
\name{use_anem_function}
\alias{use_anem_function}
\title{Use anem function}
\usage{
use_anem_function(function_name, ...)
}
\arguments{
\item{function_name}{Name of function to use, as a string}
}
\description{
Use anem function not exported from package
}
