% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_geoprocessing.R
\name{utm_zone_to_proj4}
\alias{utm_zone_to_proj4}
\title{Convert utm zone to proj4string}
\usage{
utm_zone_to_proj4(utm_zone)
}
\arguments{
\item{utm_zone}{UTM zone as numeric integer, or string}
}
\value{
The proj4string of the UTM Zone
}
\description{
Convert utm zone to proj4string
}
\examples{
utm_zone_to_proj4(32)
}
