% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_data.R
\docType{data}
\name{wells_example}
\alias{wells_example}
\title{Wells example}
\format{wells \code{data.frame} object (needs to be supplyed to \code{define_wells})}
\usage{
wells_example
}
\description{
Wells example
}
\keyword{datasets}
