% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_geoprocessing.R
\name{bounds_to_sf}
\alias{bounds_to_sf}
\title{Get bounds as sf object}
\usage{
bounds_to_sf(bounds, crs)
}
\description{
Convert bounds object with bID, x1, y1, x2, y2 to sf object
}
\examples{
bounds <- data.frame(bID=1,x1=1,y1=1,x2=3,y2=3)
#bounds_to_sf(bounds,crs=4326)
}
\keyword{internal}
