% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_geoprocessing.R
\name{gen_circleFun}
\alias{gen_circleFun}
\title{Generate outline of circle}
\usage{
gen_circleFun(circle, npoints = 100)
}
\arguments{
\item{circle}{A \code{data.frame} with $x, $y, and $r (or $R) columns. $id optional}
}
\description{
Generate outline of circle
}
\keyword{internal}
