% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_imaging.R
\name{generate_image_wells}
\alias{generate_image_wells}
\title{Generate image wells across rectangular boundaries}
\usage{
generate_image_wells(wells, aquifer, include_image_columns = FALSE)
}
\arguments{
\item{wells}{Wells with columns Q, R, diam, x, y}

\item{aquifer}{Aquifer object with bounds defined including columns for bID, bound_type, m, b}

\item{include_image_columns}{If FALSE, no new columns are added. If TRUE, columns for well
imaging are included in the result.}
}
\value{
A tibble containing original and mirrored wells, with the following columns:
  x, y, Q, diam, path, orig_wID, transform (boundary type), source_bound.
  Columns in wells that are not reproduced by this function are filled with \code{NA}.
}
\description{
Mirror wells across 2 parallel bounds, then the two perpendicular bounds
}
\section{Method}{

The original wells are labelled L0. These wells are mirrored across a set of
parallel boundaries. Each mirrored well (L1) has a source well (L0) and source
boundary (the bound across which it was mirrored). The L1 wells then need to
be mirrored across the other parallel boundary. The wells from level Lx are mirrored
across the other boundary (ie, not the source boundary), and each mirror
well (L+) has a source well (Lx) and source boundary (the bound across
which it was mirrored). This is done for a first set of parallel boundaries,
then repeated for the other two perpendicular boundaries.
}

\section{Notes}{

The function requires that the wells are labeled with a column of identifiers, wID. If it is not
present, the function generates them. The image well wID's are always generated automatically.
not present.
}

\examples{
well1 <- define_wells(x=50,y=50,Q=20,R=100,diam=1)
well2 <- define_wells(x=25,y=75,Q=20,R=100,diam=1)
wells <- define_wells(bind_rows(well1,well2))
bounds <- data.frame(bound_type=c("CH","NF","NF","NF"),m=c(Inf,0,Inf,0),b=c(0,0,100,100)) \%>\% define_bounds()
aquifer <- define_aquifer("unconfined",Ksat=1e-4,bounds=bounds)
image_wells <- generate_image_wells(wells,bounds)

bounds <- data.frame(bound_type=c("PB","PB","CH","NF"),m=c(Inf,0,Inf,0),b=c(0,0,100,100)) \%>\% define_bounds()
aquifer <- define_aquifer("unconfined",Ksat=1e-4,bounds=bounds)
image_wells <- generate_image_wells(well1,bounds,include_image_columns=TRUE)

ggplot() +
  geom_segment(data=bounds,aes(x1,y1,xend=x2,yend=y2)) +
  geom_point(data=image_wells,aes(x,y,color=as.factor(orig_wID))) +
  scale_shape_manual(values=c(1,16)) +
  coord_equal()

tstart <- Sys.time()
myfunc <- function(N,group_imaging) {
  for (i in 1:N) {
    generate_image_wells(wells,bounds,group_imaging=group_imaging)
  }
}
system.time(myfunc(5,TRUE))
system.time(myfunc(5,FALSE))

wells <- define_wells(x=rnorm(20),y=rnorm(20),Q=20,R=100,diam=1)
system.time(generate_image_wells(wells,bounds,group_imaging=TRUE))
system.time(generate_image_wells(wells,bounds,group_imaging=FALSE))
}
