% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_potentials.R
\name{get_bounds_behavior}
\alias{get_bounds_behavior}
\title{Get hydraulic behavior on bounds}
\usage{
get_bounds_behavior(wells, bounds, h0, Ksat, z0 = z0, aquifer_type,
  length.out = 100)
}
\arguments{
\item{wells}{wells object with each row containing rate Q [m^3/s], diam [m],
radius of influence roi [m], & coordinates x0 [m], y0 [m]}

\item{h0}{resting hydraulic head [m] of the aquifer without any pumping. For
an unconfined aquifer, this must be the thickness of the water table.}

\item{Ksat}{saturated hydraulic conductivity [m/s]}

\item{z0}{thickness of the confined aquifer (confining case only)}

\item{aquifer_type}{"confined" or "unconfined". Determines the calculation
and output units.}

\item{length.out}{The number of points to check on each boundary}
}
\value{
Returns a \code{data.frame}, with \code{length.out} rows for each bID.
Each row represents a point along bID and contains the head and flow as:
 dx = -Ksat * dh/dx, and
 dy = -Ksat * dh/dy
}
\description{
Get hydraulic head and flow on boundaries
}
