% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_geoprocessing.R
\name{get_contourlines}
\alias{get_contourlines}
\title{Get contour lines}
\usage{
get_contourlines(df = NULL, nlevels = 10, drop_outer = TRUE,
  levels = NULL, ..., type = "data.frame", crs = 4326)
}
\arguments{
\item{df}{A data.frame containing x, y, and z columns. x and y must form a raster,
meaning every x must be represented for each y, and vice versa.}

\item{nlevels}{An integer containing number of levels to return. Used IFF levels not supplied.}

\item{drop_outer}{If \code{TRUE}, \code{nlevels+2} will be calculated, dropping the outer 2 values. Used IFF levels not supplied.}

\item{levels}{A vector of values at which contours should be plotted. If used, nlevels is ignored.}

\item{...}{Should contain x, y, and z, and then df will be ignored.}

\item{type}{Should be either "data.frame" or "sf"}

\item{crs}{The target crs for "sf" output. Used as input in \code{sf::st_as_sf}}
}
\description{
Get contour lines, using a wrapper for contourLines
}
\examples{
# Example 1
df <- expand.grid(x=-10:10,y=-10:10)
df$z <- df$x^2
cl <- get_contourlines(df,nlevels=5)
unique(cl$level)

library(ggplot2)
ggplot() +
  geom_raster(data=df,aes(x,y,fill=z)) +
  geom_path(data=cl,aes(x,y,group=line))

cl <- get_contourlines(df,levels=c(15,20,60))
unique(cl$level)

# Example 2, with "sf" object
library(sf)
df <- expand.grid(x=seq(-5,5,length.out=20),y=seq(-5,5,length.out=20))
df$z <- sqrt(df$x^2+df$y^2)

cl <- get_contourlines(df,levels=seq(1,10,by=1.5), type="sf")
ggplot() +
  geom_raster(data=df,aes(x,y,fill=z)) +
  geom_sf(data=cl,aes())
}
