% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_geoprocessing.R
\name{get_distance_to_bounds}
\alias{get_distance_to_bounds}
\title{Get distance from point to bounds}
\usage{
get_distance_to_bounds(loc, bounds, return_locations = FALSE)
}
\arguments{
\item{loc}{Location given as c(x,y) or as data.frame and $x, $y}

\item{bounds}{Boundary object with m and b or x1, y1, x2, y2}
}
\description{
Get distance from point to bounds. Location on the boundaries
has to fall within the segment x1, y1, x2, y2
}
\examples{
\dontrun{
bounds <- define_bounds(bounds_example)
loc <- c(150,150)
get_distance_to_bounds(loc,bounds)

bounds_df <- data.frame(bound_type=c("CH","NF","NF","NF"),m=c(-2,0.5,-2,0.5),b=c(0,0,100,20))
bounds <- define_bounds(bounds_df)
loc <- data.frame(x=c(-200,0,200,50),y=c(-200,0,200,-50))
get_distance_to_bounds(loc,bounds)

library(ggplot2)
ggplot(bounds) +
  geom_segment(data=bounds,aes(x1,y1,xend=x2,yend=y2)) +
  geom_point(data=loc,aes(x,y)) +
  coord_equal()
}
}
\keyword{internal}
