% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_potentials.R
\name{get_hydraulic_head}
\alias{get_hydraulic_head}
\title{Get cumulative effect of wells at location}
\usage{
get_hydraulic_head(loc, wells, aquifer)
}
\arguments{
\item{loc}{coordinates vector as c(x,y), with units of [m] or as data.frame with columns $x and $y}

\item{wells}{wells object with each row containing rate Q [m^3/s], diam [m],
radius of influence R [m], & coordinates x [m], y [m]}

\item{aquifer}{Afuifer object containing aquifer_type, h0, Ksat, bounds, z0 (for confined case only)}
}
\value{
The output is the hydraulic head \code{loc}, accounting for the
  cumulative effect of all \code{wells} (dP), which is given as hydraulic
  head [units=m] if \code{aquifer_type="confined"} or discharge potential
  [m^2] if \code{aquifer_type="unconfined"}. Then the head at \code{loc} is:

  \describe{
  \item{aquifer_type="confined"}{\eqn{h=h_0+dP}}
  \item{aquifer_type="unconfined"}{\eqn{h=\sqrt{h_0^2+dP}}} }
}
\description{
Get the hydraulic head at a location, accounting for the aquifer type
(confined or unconfined), resting head, and cumulative effect of all wells. This function
is a wrapper around \code{get_potential_differential()} to get the actual hydraulic head instead of the differential potential.
}
\examples{
well1 <- define_wells(x=0,y=0,Q=1e-3,diam=0.75,R=300)
well2 <- define_wells(x=0.5,y=0.25,Q=-2e-3,diam=0.8,R=300)
aquifer <- define_aquifer(aquifer_type="confined",Ksat=0.00001,h0=0,z0=30)


get_hydraulic_head(well1,loc=c(5,5),aquifer)
get_hydraulic_head(well2,loc=c(5,5),aquifer)
wells <- rbind(well1,well2)

get_hydraulic_head(wells,loc=c(5,5),aquifer=aquifer)

grid_pts <- expand.grid(x=seq(0,10,by=5),y=seq(0,10,by=5))
aquifer_unconfined <- define_aquifer(aquifer_type="confined",Ksat=0.00001,h0=0,z0=30)
get_hydraulic_head(well1,loc=grid_pts,aquifer=aquifer_unconfined)
get_hydraulic_head(well2,loc=grid_pts,aquifer=aquifer_unconfined)
get_hydraulic_head(wells,loc=grid_pts,aquifer=aquifer_unconfined)

# Ensure potentials are even when pumping is symmetric (and opposite sign)
well1 <- define_wells(x=0,y=0,Q=1e-3,diam=0.5,R=300)
well2 <- define_wells(x=1,y=1,Q=-2e-3,diam=0.5,R=300)
well3 <- define_wells(x=0,y=0.5,Q=2e-3,diam=0.5,R=300)
well4 <- define_wells(x=0.5,y=1,Q=-2e-3,diam=0.5,R=300)
wells_even <- rbind(well1,well1,well2,well3,well4)
grid_pts_even <- data.frame(x=c(0,0.5,1),y=c(1,0.5,0))
get_hydraulic_head(wells_even,loc=grid_pts_even,aquifer=aquifer_unconfined)
}
