% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_imaging.R
\name{get_mirror_point}
\alias{get_mirror_point}
\title{Mirror a well across a boundary}
\usage{
get_mirror_point(well, boundary, new_wID = NA)
}
\arguments{
\item{well}{single well as a list (or sf object single well feature)
containing coordinates (well$x, well$y), and pumping rate (well$Q),
which is positive for injection, negative for pumping. Should also have:
diam, path, origin.}

\item{boundary}{single line as a list (or sf object single line feature)
containing slope (as boundary$m) and intercept (as boundary$b), as well as
the boundary type (bound_type) as "CH" for constant head or "NF" for no flow}

\item{new_wID}{integer ID for the new well to be created}
}
\value{
a row (same class as well input) containing the new imaged well and
  the following columns: x, y, Q, diam, path, origin, transform
  (boundary type),
  source_bound
}
\description{
Mirror a well across a boundary, specified as constant head (CH) or no flow
(NF) and defined by the line \eqn{m x + b})
}
\examples{
well <- define_wells(wID=1,x=0,y=0,Q=0.5,diam=0.75,path=1,origin=1,well_group="a")
boundary <- data.frame(bID=1,m=-1,b=1,bound_type="NF")
get_mirror_point(well, boundary, new_wID=2)

well <- define_wells(wID=1,x=2,y=2,Q=0.5,diam=0.75,path=1,origin=1,well_group="a")
boundary <- data.frame(bID=1,m=Inf,b=1,bound_type="NF")
get_mirror_point(well, boundary, new_wID=2) \%>\% select(x,y)


well <- define_wells(wID=1,x=2,y=2,Q=0.5,diam=0.75,path=1,origin=1,well_group="a")
boundary <- data.frame(bID=1,m=0,b=1,bound_type="NF")
get_mirror_point(well, boundary, new_wID=2) \%>\% select(x,y)
}
