% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_assess.R
\name{get_segments_behavior}
\alias{get_segments_behavior}
\title{Get hydraulic behavior along segments}
\usage{
get_segments_behavior(wells, segments, aquifer, length.out = 100,
  eps = 1e-10)
}
\arguments{
\item{wells}{wells object with each row containing rate Q [m^3/s], diam [m],
radius of influence R [m], & coordinates x [m], y [m]}

\item{segments}{A data.frame containing, m, b, sID, x_1, x_2, y_1, y_2}

\item{length.out}{The number of points to check on each segment}
}
\value{
Returns a \code{data.frame}, with \code{length.out} rows for each bID.
Each row represents a point along sID and contains the head and flow as:
 dx = -Ksat * dh/dx, and
 dy = -Ksat * dh/dy
}
\description{
Get hydraulic head and flow on segments
}
\examples{
wells <- define_wells(x=c(50,10),y=c(25,2.5),Q=c(0.5,-0.2),diam=c(0.05,0.08)) \%>\%
  mutate(R=get_ROI(Ksat=0.0000005,h=10,t=630720000,n=0.5,method="aravin-numerov"))  # 630720000 - 20 years
bounds_df <- tibble(bound_type=c("CH","NF","NF","NF"), m=c(0.5,-2,0.5,-2),b=c(0,100,20,50), bID=as.numeric(1:4))
aquifer <- define_aquifer("unconfined",1e-4,bounds=bounds_df,h0=100)
well_images <- generate_image_wells(wells,bounds,num_levels=2) \%>\% filter(max_mirror_dist<R)
segments <- bounds[1:2,] \%>\% rename(sID=bID) \%>\% mutate(x_1=c(0,50),x_2=c(50,75),y_1=c(0,50),y_2=c(50,25)) \%>\%
  dplyr::mutate(x_unit_norm=sqrt(m^2/(1+m^2)),y_unit_norm=-sign(m)*sqrt(1/(1+m^2)))
segments_behavior <- get_segments_behavior(well_images,segments,aquifer) \%>\% as_tibble()
p_domain <- ggplot() +
  geom_point(data=well_images,aes(x0,y0,fill=rate),color="black",size=2,shape=21) +
  scale_fill_gradient2(low="blue",high="red",mid="gray")+
  geom_abline(data=bounds,aes(slope=m,intercept=b,linetype=bound_type,color=as.factor(bID))) +
  geom_segment(data=segments,aes(x_1,y_1,xend=x_2,yend=y_2),linetype="dashed") +
  coord_equal()
p_segments <- ggplot(segments_behavior \%>\% gather(var,val,head,flow_normal)) +
  geom_line(aes(dist,val,color=as.factor(sID))) + facet_grid(var~sID,scales="free")
gridExtra::grid.arrange(p_domain,p_segments,nrow=1)
}
