% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_imaging.R
\name{mirror_across_bounds}
\alias{mirror_across_bounds}
\title{Recursively mirror wells across boundaries}
\usage{
mirror_across_bounds(wells, bounds, num_levels, first_mirror = TRUE)
}
\arguments{
\item{wells}{Wells with columns Q, diam, x, y}

\item{bounds}{Bounds with columns bID, bound_type, m, b}

\item{num_levels}{Maximum number of iterations to image wells (and well images). Greater than or equal to 1.}
}
\value{
A data.frame containing original and mirrored wells, with the following columns:
  x, y, Q, diam, path, origin, transform (boundary type), source_bound.
  Columns in wells that are not reproduced by this function are filled with \code{NA}.
}
\description{
Recursively mirror wells across boundaries
}
\section{Method}{

The original wells are labelled L0. These wells are mirrored across the
boundaries, and each mirrored well (L1) has a source well (L0) and source
boundary (the bound across which it was mirrored) The L1 wells then need to
be mirrored across all boundaries except for the source boundary (which
would replicate the actual well). The wells from level Lx are mirrored
across the boundaries (except for the source boundary), and each mirror
well (L+) has a source well (Lx) and source boundary (the bound across
which it was mirrored).
}

\section{Notes}{

The function requires that the wells are labeled with a column of identifiers, wID. If it is not
present, the function generates them. The image well wID's are always generated automatically.
not present.
}

\examples{
wells <- data.frame(x=c(0,0.5),y=c(0,0.25),Q=c(0.5,-0.2),diam=c(0.75,0.8))
bounds <- data.frame(m=c(1,-1),b=c(0.5,1),bound_type=c("CH","NF"),bID=c(1,2))
mirror_across_bounds(wells,bounds,num_levels=1)
mirror_across_bounds(wells,bounds,num_levels=2)
}
