% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_potentials.R
\name{plot_bounds_behavior}
\alias{plot_bounds_behavior}
\title{Plot behavior on boundaries}
\usage{
plot_bounds_behavior(wells_noimages, well_images, bounds, h0, Ksat,
  z0 = NA, aquifer_type, length.out = 100)
}
\arguments{
\item{h0}{resting hydraulic head [m] of the aquifer without any pumping. For
an unconfined aquifer, this must be the thickness of the water table.}

\item{Ksat}{saturated hydraulic conductivity [m/s]}

\item{z0}{thickness of the confined aquifer (confining case only)}

\item{aquifer_type}{"confined" or "unconfined". Determines the calculation
and output units.}

\item{length.out}{The number of points to check on each boundary}
}
\value{
Two ggplot objects that show behavior on the boundaries, one for head and one for flow
}
\description{
Plot behavior on boundaries with and without well images
}
\examples{
wells_noimages <- data.frame(x0=c(50,5),y0=c(25,2.5),rate=c(0.5,-0.2),diam=c(0.05,0.08)) \%>\% as_tibble() \%>\%
   mutate(roi=get_radius_of_influence(Ksat=0.0000005,h=10,t=630720000,n=0.5,method="aravin-numerov"))  # 630720000 - 20 years
bounds <- tibble(bound_type=c("CH","NF","NF","NF"),
                 m=c(0.8,-1.25,0.8,-1.25),b=c(3,100,-25,1),
                 bID=as.numeric(1:4),bGroup=c(2,1,2,1))
well_images <- generate_image_wells(wells_noimages,bounds,num_levels=6) \%>\%
   filter(max_mirror_dist<roi)
ggplot() + geom_point(data=well_images,aes(x0,y0)) +
  geom_abline(data=bounds,aes(slope=m,intercept=b)) + coord_equal()
gg_list <- plot_bounds_behavior(wells_noimages,image_wells,bounds,h0=1000,Ksat=1,z0=10?0,aquifer_type="confined",length.out=50)
}
