% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_printing.R
\name{print.aquifer}
\alias{print.aquifer}
\title{S3 Print method for class "aquifer"}
\usage{
\method{print}{aquifer}(aquifer)
}
\arguments{
\item{aquifer}{Object of type aquifer}
}
\description{
S3 Print method for class "aquifer"
}
