% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anem_imaging.R
\name{reconstruct_image_pumping}
\alias{reconstruct_image_pumping}
\title{Reproduce pumping at image wells}
\usage{
reconstruct_image_pumping(image_wells)
}
\arguments{
\item{wells}{A data.frame object of wells, containing well_image column with
"Actual", "Image (+Q)", or "Image (-Q)" text.}
}
\value{
A tibble containing wells where Image wells contain pumping rates similar to
generate_image_wells.
}
\description{
Reproduce pumping of actual wells at image wells
}
\examples{
well1 <- define_wells(x=50,y=50,Q=5,R=100,diam=1)
well2 <- define_wells(x=25,y=75,Q=-2,R=100,diam=1)
wells <- define_wells(bind_rows(well1,well2))
bounds <- data.frame(bound_type=c("CH","NF","NF","NF"),m=c(Inf,0,Inf,0),b=c(0,0,100,100)) \%>\% define_bounds()
aquifer <- define_aquifer("unconfined",Ksat=1e-4,bounds=bounds)
image_wells <- generate_image_wells(wells,bounds)
image_wells_image_NA <- image_wells \%>\%
  dplyr::mutate(Q=dplyr::case_when(grepl("Image",well_image)~as.numeric(NA),TRUE~Q))
image_wells_reconstructed <- reconstruct_image_pumping(image_wells_image_NA)
identical(image_wells,image_wells_reconstructed)
}
